App Security
App Security

+++ Apps/Text/startSub.php
@@ -1,4 +1,4 @@
-<?php include_once "../../wd_protect.php";
+<?php if(is_file("../../wd_protect.php")){ include_once "../../wd_protect.php"; }
     $title = test_input($_POST["title"]);
     $text = htmlspecialchars_decode($_POST["text"], ENT_QUOTES);
     file_put_contents($wd_file . $title, $text);
